Code restructured
