+++ src/components/Header/Header.jsx
@@ -1,0 +1,23 @@
+import reactImage from "../../assets/react-core-concepts.png";
+import "./Header.css";
+
+const reactDescription = ["Fundamental", "Crucial", "Core"];
+
+function genRandomInt(max) {
+  return Math.floor(Math.random() * (max + 1));
+}
+
+export default function Header() {
+  const description = reactDescription[genRandomInt(2)];
+
+  return (
+    <header>
+      <img src={reactImage} alt="Stylized atom" />
+      <h1>React Essentials</h1>
+      <p>
+        {description} React concepts you will need for almost any app you are
+        going to build!
+      </p>
+    </header>
+  );
+}
